Added Qr code. Renamed Resolve case study. Removed some tech from mobile developement
